Move extract_usage() from Configuration to Service
